contrib(publish): clean up publishing tool
