+++ contrib/inertia-publish/main.go
@@ -1,11 +1,11 @@
 /*
 
-Inertia-publish is a tool for helping publish Inertia releases to various distribution
-channels.
+Inertia-publish is a tool for helping publish Inertia releases to various Launch Pad
+distribution channels.
 
-For example, to generate completions for zsh:
+For example, to publish to the Launch Pad Hombrew Tap (https://github.com/ubclaunchpad/homebrew-tap):
 
-	inertia contrib publish [channel]
+	inertia contrib publish homebrew
 
 Learn more about `inertia/contrib` tools:
 
@@ -16,14 +16,8 @@ package main
 
 import (
 	"bytes"
-	"crypto/sha256"
-	"encoding/hex"
-	"fmt"
-	"io/ioutil"
 	"os"
-	"os/exec"
 	"path/filepath"
-	"strings"
 	"text/template"
 
 	"github.com/spf13/cobra"
@@ -32,98 +26,6 @@ import (
 )
 
 //go:generate go run github.com/UnnoTed/fileb0x b0x.yml
-
-// TemplateVariables carries variables available to templates
-type TemplateVariables struct {
-	Version string
-	Sha256  map[string]string
-}
-
-func generateSums(buildDir string) map[string]string {
-	sums := make(map[string]string)
-	if err := filepath.Walk(buildDir, func(path string, info os.FileInfo, err error) error {
-		parts := strings.Split(info.Name(), ".")
-		if info.IsDir() || parts[0] != "inertia" {
-			return nil
-		}
-		data, err := ioutil.ReadFile(path)
-		if err != nil {
-			return err
-		}
-		sum := sha256.Sum256(data)
-		build := strings.Join(parts[len(parts)-2:], ".")
-		sums[build] = hex.EncodeToString(sum[:])
-		return nil
-	}); err != nil {
-		out.Fatal(err)
-	}
-
-	// if no sums were generated, some setup is likely missing
-	if len(sums) == 0 {
-		out.Fatal("no binary sums generated - was '.scripts/build_release.sh' run?")
-	}
-
-	return sums
-}
-
-func newCommand(dir, name string, args ...string) *exec.Cmd {
-	cmd := exec.Command(name, args...)
-	cmd.Dir = dir
-	cmd.Stderr = os.Stderr
-	cmd.Stdout = os.Stdout
-	return cmd
-}
-
-func getCloneURL(repository string) string {
-	// GIT_CREDENTIALS should be $USER:$TOKEN
-	if creds := os.Getenv("GIT_CREDENTIALS"); creds != "" {
-		println("Using GIT_CREDENTIALS to clone")
-		// "Personal access tokens can only be used for HTTPS Git operations.""
-		// https://docs.github.com/en/free-pro-team@latest/github/authenticating-to-github/creating-a-personal-access-token#using-a-token-on-the-command-line
-		return fmt.Sprintf("https://%s@github.com/%s.git", creds, repository)
-	}
-	return fmt.Sprintf("https://github.com/%s.git", repository)
-}
-
-type channelConfig struct {
-	Repository string
-	Template   string
-	PublishAs  string
-}
-
-func (c *channelConfig) Publish(version, outdir string, renderedTemplate []byte, dryRun bool) error {
-	// load repository
-	repoPath := filepath.Join(outdir, c.Repository)
-	os.RemoveAll(repoPath)
-	clone := newCommand(outdir, "git", "clone", getCloneURL(c.Repository),
-		// clone into subdirectory
-		c.Repository)
-	if err := clone.Run(); err != nil {
-		return err
-	}
-
-	// update file
-	publishFile := filepath.Join(repoPath, c.PublishAs)
-	os.Remove(publishFile)
-	if err := ioutil.WriteFile(publishFile, renderedTemplate, os.ModePerm); err != nil {
-		return err
-	}
-
-	// commit changes
-	commit := newCommand(repoPath, "git", "commit", "-a",
-		"-m", fmt.Sprintf("inertia-publish: release %s", version))
-	if err := commit.Run(); err != nil {
-		return err
-	}
-
-	// publish
-	if dryRun {
-		out.Println("dryrun enabled - aborting")
-		return nil
-	}
-	publish := newCommand(repoPath, "git", "push", "origin", "HEAD")
-	return publish.Run()
-}
 
 var channelConfigs = map[string]channelConfig{
 	"homebrew": {
@@ -149,13 +51,16 @@ func main() {
 		Args:   cobra.MinimumNArgs(1),
 		Short:  "Publish Inertia releases to various distribution channels.",
 		Run: func(cmd *cobra.Command, args []string) {
+			version := args[0]
 			outdir, _ := cmd.Flags().GetString(flagOutput)
 			dryRun, _ := cmd.Flags().GetBool(flagDryRun)
+
 			os.MkdirAll(outdir, os.ModePerm)
-			data := TemplateVariables{
-				Version: args[0],
-				Sha256:  generateSums(outdir),
+			data, err := newTemplateVariables(version, outdir)
+			if err != nil {
+				out.Fatalf("failed to load release data: %s\n", err)
 			}
+
 			for _, channel := range args[1:] {
 				conf, exists := channelConfigs[channel]
 				if !exists {
